reduce tools hotspot pressure
